+++ x64barebones/Kernel/syscalls/sys_wait_for_children.c
@@ -2,7 +2,7 @@
 #include <syscalls.h>
 
 
-unsigned int sys_wait_for_children(uint64_t rsp, uint64_t ss){
+uint64_t sys_wait_for_children(uint64_t rsp, uint64_t ss){
 	wait_for_children(rsp,  ss);
 	return 1;
 }
